+++ client/src/components/editor/ProofStep.tsx
@@ -2,5 +2,5 @@ import React from "react";
 import { InteractiveGoalsWithHints } from "../../api/rpc_api";
 
 export function ProofStep({step, idx}: {step:InteractiveGoalsWithHints, idx: number}) {
-  return <div>Step {idx}</div>
+  return <div key={idx} >Step {idx}</div>
 }
