+++ client/src/components/editor/types.ts
