+++ client/src/components/level/ExercisePanel.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import { GameEditor } from "../infoview/GameEditor"
+import { GameEditor } from "../editor/GameEditor"
 
 /** The center-piece of a playable level */
 export function ExercisePanel({codeviewRef, visible=true}: {codeviewRef: React.MutableRefObject<HTMLDivElement>, visible?: boolean}) {
